Documents/fastapi_ecommerce_backend/requirements.txt
aiofiles==24.1.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.9.0
attrs==23.2.0
bcrypt==4.0.1
blinker==1.9.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
cryptography==45.0.6
cymem==2.0.11
distro==1.9.0
dnspython==2.7.0
docx2txt==0.9
ecdsa==0.19.1
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.112.2
filelock==3.19.1
Flask==3.1.0
fsspec==2025.7.0
h11==0.14.0
hf-xet==1.1.8
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.7
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.1
Mako==1.3.10
marisa-trie==1.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
numpy==1.26.4
openai==1.75.0
optional-django==0.3.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pluggy==1.6.0
preshed==3.0.10
psycopg==3.2.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
react==4.3.0
regex==2025.7.34
requests==2.32.3
rsa==4.9.1
safetensors==0.6.2
selenium==4.21.0
setuptools==80.9.0
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
srsly==2.5.1
starlette==0.38.6
striprtf==0.0.29
sympy==1.14.0
thinc==8.2.5
tokenizers==0.19.1
torch==2.4.0
tqdm==4.67.1
transformers==4.43.3
trio==0.25.1
trio-websocket==0.11.1
typer==0.9.4
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.3.4
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yt-dlp==2025.5.22
